+++ test/nodejs/serializeProject.mjs
@@ -1,21 +1,17 @@
 #!/usr/bin/env node
 
 //
-// Dumps a JClic project in JSON format
+// Dumps a JClic project in JSON format (module version)
 //
 // Usage:
 // node serializeProject.js [filename]
 //
 
-/* global process, global, require, console */
+/* global process, global, console */
 
-var fs = require('fs');
-
-// Load the main JClic module.
-// Here this is done with a relative path. In other contexts just install
-// the 'jclic' NPM package and require it, like in:
-// `var jclic = require('jclic');`
-var jclic = require('../../dist/jclic-node.js');
+import fs from 'node:fs';
+import JClicObject from '../../dist/jclic-node.js';
+const { JClicProject, $ } = JClicObject;
 
 if (process.argv.length < 3) {
   console.log('Usage: serializeProject.js {path/to/project.jclic[.json]}');
@@ -28,15 +24,15 @@ var file = process.argv[2];
 
 const isXML = file.endsWith('.jclic');
 
-const project = new jclic.JClicProject();
+const project = new JClicProject();
 const contents = fs.readFileSync(file, 'utf8');
 
 if (isXML) {
   // Read file and parse it into a DOM object
-  var doc = new global.DOMParser().parseFromString(contents, 'text/xml');
+  var doc = new global.DOMParser().parseFromString(contents, 'application/xml');
 
   // Create a JClicProject and initialize it with the file contents
-  project.setProperties(jclic.$(doc).find('JClicProject'), file, null, {});
+  project.setProperties($(doc).find('JClicProject'), file, null, {});
 }
 else {
   var doc = JSON.parse(contents);
@@ -44,3 +40,4 @@ else {
 }
 
 console.log(project.getJSON(2));
+